+++ libminimk/syscall/listen_posix.hpp
@@ -12,12 +12,13 @@
 #include <sys/socket.h> // for listen
 
 /// Testable minimk_syscall_listen implementation.
-template <decltype(minimk_syscall_clearerrno) M_minimk_syscall_clearerrno = minimk_syscall_clearerrno,
-          decltype(minimk_syscall_geterrno) M_minimk_syscall_geterrno = minimk_syscall_geterrno,
-          decltype(listen) M_sys_listen = listen>
+template <
+        decltype(minimk_syscall_clearerrno) M_minimk_syscall_clearerrno = minimk_syscall_clearerrno,
+        decltype(minimk_syscall_geterrno) M_minimk_syscall_geterrno = minimk_syscall_geterrno,
+        decltype(listen) M_sys_listen = listen>
 minimk_error_t minimk_syscall_listen_impl(minimk_syscall_socket_t sock, int backlog) noexcept {
     M_minimk_syscall_clearerrno();
-    int rv = M_sys_listen((int)sock, backlog);
+    int rv = M_sys_listen(sock, backlog);
     return (rv == 0) ? 0 : M_minimk_syscall_geterrno();
 }
